refactor(react-gui): give the animation services a folder
`animation.service.ts` (732 lines) and `animDetail.service.ts` (639) were
two halves of one feature that had grown apart: the panel's half and the
inspector's half, each with its own copy of the AnimMgr scaffolding, sharing
two files in `helpers/` that existed only for them. Eighteen calls between
them, and no way to see which was which without reading both.

`anim/` is one file per thing the animation does. `timeline.ts` lists the
strip and `transport.ts` plays it; `edit.ts` changes it; `detail.ts` is one
element as the inspector shows it, `targets.ts` what that element can point
at, and `genericProps.ts` its raw property tab. `read.ts` holds the readers
the timeline and playback share, so the two cannot describe the same manager
differently. `helpers/animResolve.ts` and `helpers/animElementType.ts` move
in as `resolve.ts` and `elementType.ts` -- nothing outside the feature ever
used them, which is what `helpers/` is supposed to mean.

`anim.service.ts` is the registry entry: the eighteen keys, and the four
playback functions that leave by a different door because the worker's frame
loop and view activation call them rather than the renderer.

No behaviour change; the existing animation tests pass with only their
import paths moved.
