+++ tritium/react-gui/src/renderer/worker/server/services/anim/resolve.ts
@@ -1,20 +1,9 @@
-/**
- * @file worker/server/services/helpers/animResolve.ts
- * @description Shared worker-side helpers for the animation services
- * (`animation.service.ts`, `animDetail.service.ts`): scene/AnimMgr resolution,
- * TimeValue construction, and safe wrapper reads.
- *
- * Lifted here so both anim services use one implementation. Wrapper calls are
- * synchronous in the Web Worker thread.
- */
-
 import type { Scene } from "@cuemol/core/src/wrappers/Scene";
 import type { AnimMgr } from "@cuemol/core/src/wrappers/AnimMgr";
 import type { AnimObj } from "@cuemol/core/src/wrappers/AnimObj";
 import type { TimeValue } from "@cuemol/core/src/wrappers/TimeValue";
 import type { WorkerContext } from "@renderer/worker/server/types/WorkerContext";
-import { getSceneOrNull } from "./sceneResolver";
-
+import { getSceneOrNull } from "../helpers/sceneResolver";
 // --- safe wrapper reads (a getter may throw for missing-on-subclass cases) ---
 
 export function safeNum(read: () => number): number {
